tests/wpt/web-platform-tests/css/css-flexbox-1/getcomputedstyle/flexbox_computedstyle_flex-flow-row-nowrap.html

<!DOCTYPE html>
<title>flexbox | computed style | flex-flow: row nowrap</title>
<link rel="author" href="http://opera.com" title="Opera Software">
<link rel="help" href="http://www.w3.org/TR/css-flexbox-1/#propdef-flex-flow">
<link rel="help" href="http://www.w3.org/TR/css-flexbox-1/#valdef-flex-direction-row">
<link rel="help" href="http://www.w3.org/TR/css-flexbox-1/#valdef-flex-wrap-nowrap">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="flags" content="dom">
<style>
body {
	color: red;
	flex-flow: row nowrap;
}
</style>
<div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("flex-direction"),
	          "row");
    assert_equals(getComputedStyle(body).getPropertyValue("flex-wrap"),
	          "nowrap");
});
</script>
